+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 0.3
 - Prevented replacement of built-in types.
+- added property to target to support adding '.generated' extension to files.
 
 ## Version 0.2
 - Fixed bug in entity syntax, added 'members' property.